Restructured a project and added readme
