+++ API/Models/account_info.js
@@ -1,6 +1,6 @@
-var mongoose = require('mongoose');
-var Schema = mongoose.Schema;
-var account_info = new Schema({
+const mongoose = require('mongoose');
+const Schema = mongoose.Schema;
+const account_info = new Schema({
 savings:{
 	type:String,
 	required:false
