+++ README.md
@@ -1,0 +1,10 @@
+# Banking chatbot
+
+This bot will provide banking details upon users query. like balance, last 5 transaction details, cibil score.
+
+## Getting Started
+
+1. Install NodeJs (version 8.12.0 or above) & npm (version 6.4.1 or above)
+2. Clone the project. `git clone https://github.com/Pirimid/chatbot.git`
+3. Install node dependency. `npm install`
+4. To run the node server in devlopment. `npm start`
