+++ index.js
@@ -1,5 +1,5 @@
 'use strict';
-var express  = require('express'),
+const express  = require('express'),
 bodyParser   = require('body-parser'),
 http         = require('http'),
 config       = require('./Config'),
@@ -11,7 +11,7 @@ mongoose.Promise = global.Promise;
 mongoose.connect(config.dbUrl);
 server.use(bodyParser.urlencoded({ extended: true }));
 server.use(bodyParser.json());
-var routes = require('./API/Routes/Routes'); //importing route
+const routes = require('./API/Routes/Routes'); //importing route
 routes(server); //register the route
 server.listen((process.env.PORT || 8000), function () {
     console.log("Server is up and listening on port" + process.env.PORT);
